Only change incomingID for policies
If an object that has an anchor also has an id key, mt will
treat it like a policy and change the incomingID, but only when
processing anchors, not aliases. This leads to a mismatch in
anchors and aliases.

This commit ensures that incomingID is only changed when an object with
the !policy tag is encountered.

Related: conjurinc/ops#738